Change from Accordion to Table
Created a `ExpandableTableRow` which displays a separate table below for additional appointment details - much prettier IMHO.
Inspired by https://codesandbox.io/s/expandable-table-row-material-ui-tr3ut?file=/demo.js
