+++ firebase/src/components/ScienceClub/Invoicing/InvoiceStatusPage/index.js
@@ -1,11 +1,11 @@
 import React, { useState } from 'react'
 import { withRouter } from 'react-router-dom'
 import queryString from 'query-string'
-import { makeStyles, CssBaseline, AppBar, Toolbar, IconButton, Typography } from '@material-ui/core'
+import { makeStyles, CssBaseline, AppBar, Toolbar, IconButton, Typography, Table, TableHead, TableRow, TableCell, TableBody } from '@material-ui/core'
 import ArrowBackIcon from '@material-ui/icons/ArrowBack'
 
 import useFetchAppointments from '../../../Hooks/UseFetchAppointments'
-import AppointmentExpansionPanel from './AppointmentExpansionPanel'
+import ExpandableTableRow from './ExpandableTableRow'
 import SkeletonRows from '../../../Shared/SkeletonRows'
 import useWindowDimensions from '../../../Hooks/UseWindowDimensions'
 
@@ -16,26 +16,28 @@ const ScienceClubInvoicingStatus = props => {
     const { height } = useWindowDimensions()
 
     const [loading, setLoading] = useState(true)
-    const [expanded, setExpanded] = useState(false)
 
     const queries = queryString.parse(props.location.search)
     const appointmentTypeID = queries.appointmentTypeId
     const calendarID = queries.calendarId
     const classID = parseInt(queries.classId)
 
+    const sortByParentName = (a, b) => {
+        const aName = a.firstName
+        const bName = b.firstName
+        return (aName < bName) ? -1 : (aName > bName) ? 1 : 0;
+    }
+
     const appointments = useFetchAppointments({
         setLoading,
         appointmentTypeID,
         calendarID,
-        classID
+        classID,
+        sorter: sortByParentName
     })
 
     const navigateBack = () => {
         props.history.goBack()
-    }
-
-    const handleAppointmentSelectionChange = panel => (_, isExpanded) => {
-        setExpanded(isExpanded ? panel : false)
     }
 
     return (
@@ -51,14 +53,21 @@ const ScienceClubInvoicingStatus = props => {
                     </Typography>
                 </Toolbar>
             </AppBar>
-            {appointments !== null && appointments.map(appointment => (
-                <AppointmentExpansionPanel
-                    key={appointment.id}
-                    appointment={appointment}
-                    onAppointmentSelectionChange={handleAppointmentSelectionChange}
-                    expanded={expanded}
-                />
-            ))}
+            <Table>
+                <TableHead>
+                    <TableRow className={classes.headerRow}>
+                        <TableCell width="5%" />
+                        <TableCell width="45%" className={classes.parentNameCell}>Parent Name</TableCell>
+                        <TableCell width="25%">Invoice Status</TableCell>
+                        <TableCell width="25%">Action</TableCell>
+                    </TableRow>
+                </TableHead>
+                <TableBody>
+                    {appointments !== null && appointments.map(appointment => (
+                        <ExpandableTableRow key={appointment.id} appointment={appointment} />
+                    ))}
+                </TableBody>
+            </Table>
             {loading && <SkeletonRows rowCount={(height - 64) / 64} />}
         </div>
     )
@@ -68,6 +77,14 @@ const useStyles = makeStyles({
     main: {
         position: 'absolute',
         top: 0, right: 0, bottom: 0, left: 0
+    },
+    headerRow: {
+        '& th': {
+            textAlign: 'center'
+        }
+    },
+    parentNameCell: {
+        textAlign: 'left !important'
     }
 })
